Add useAutomationPrimaryButton helper function
