+++ features/aave/manage/sidebars/AutoBuySidebarAaveVault.tsx
@@ -15,6 +15,7 @@ import { RemoveTriggerInfoSection } from 'features/aave/components'
 import type { BuyInfoSectionProps } from 'features/aave/components/AutoBuyInfoSection'
 import { AutoBuyInfoSection } from 'features/aave/components/AutoBuyInfoSection'
 import { mapErrorsToErrorVaults, mapWarningsToWarningVaults } from 'features/aave/helpers'
+import { useAutomationPrimaryButton } from 'features/aave/manage/helpers/use-automation-primary-button'
 import { getTriggerExecutionPrice } from 'features/aave/manage/services/calculations'
 import type {
   AutoBuyTriggerAaveContext,
@@ -43,7 +44,7 @@ type AutoBuyTriggerAaveContextWithPosition = AutoBuyTriggerAaveContext & {
   position: PositionLike
 }
 
-interface AutoBuySidebarAaveVaultProps {
+export interface AutoBuySidebarAaveVaultProps {
   strategy: IStrategyConfig
   state: AutoBuyTriggerAaveContextWithPosition
   isStateMatch: (state: BasicAutomationAaveState) => boolean
@@ -340,91 +341,6 @@ export function SideBarContent(props: AutoBuySidebarAaveVaultProps) {
   return <></>
 }
 
-export function usePrimaryButton(
-  props: AutoBuySidebarAaveVaultProps,
-): SidebarSectionFooterButtonSettings {
-  const { isStateMatch, canTransitWith } = props
-  const { t } = useTranslation()
-  const editingLabel =
-    props.state.action === TriggerAction.Add
-      ? t('automation.add-trigger', {
-          feature: t(sidebarAutomationFeatureCopyMap[props.state.feature]),
-        })
-      : t('automation.update-trigger', {
-          feature: t(sidebarAutomationFeatureCopyMap[props.state.feature]),
-        })
-
-  switch (true) {
-    case isStateMatch('idle'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {},
-        disabled: true,
-        label: editingLabel,
-        steps: [1, 3],
-      }
-    case isStateMatch('editing'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {
-          props.updateState({ type: 'REVIEW_TRANSACTION' })
-        },
-        disabled: !canTransitWith({ type: 'REVIEW_TRANSACTION' }),
-        label: editingLabel,
-        steps: [1, 3],
-      }
-    case isStateMatch('review'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {
-          props.updateState({ type: 'START_TRANSACTION' })
-        },
-        disabled: !canTransitWith({ type: 'START_TRANSACTION' }),
-        label: props.state.retryCount > 0 ? t('retry') : t('protection.confirm'),
-        steps: [2, 3],
-      }
-    case isStateMatch('remove'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {
-          props.updateState({ type: 'START_TRANSACTION' })
-        },
-        disabled: !canTransitWith({ type: 'START_TRANSACTION' }),
-        label:
-          props.state.retryCount > 0
-            ? t('retry')
-            : t('automation.cancel-trigger', {
-                feature: t(sidebarAutomationFeatureCopyMap[props.state.feature]),
-              }),
-      }
-    case isStateMatch('tx'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {},
-        disabled: true,
-        label: t('automation.setting', {
-          feature: t(sidebarAutomationFeatureCopyMap[props.state.feature]),
-        }),
-        steps: [3, 3],
-      }
-    case isStateMatch('txDone'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {
-          props.updateState({ type: 'RESET' })
-        },
-        disabled: false,
-        label: t('finished'),
-      }
-  }
-  return {
-    isLoading: props.state.isLoading,
-    action: () => {},
-    disabled: true,
-    label: '',
-  }
-}
-
 export function useTextButton(
   props: AutoBuySidebarAaveVaultProps,
 ): SidebarSectionFooterButtonSettings | undefined {
@@ -462,7 +378,7 @@ export function AutoBuySidebarAaveVault(props: AutoBuySidebarAaveVaultProps) {
 
   const { strategy } = props
 
-  const primaryButton = usePrimaryButton(props)
+  const primaryButton = useAutomationPrimaryButton(props)
   const textButton = useTextButton(props)
 
   return (
