+++ features/aave/manage/sidebars/AutoSellSidebarAaveVault.tsx
@@ -14,6 +14,7 @@ import { RemoveTriggerInfoSection, type RemoveTriggerSectionProps } from 'featur
 import type { AutoSellInfoSectionProps } from 'features/aave/components/AutoSellInfoSection'
 import { AutoSellInfoSection } from 'features/aave/components/AutoSellInfoSection'
 import { mapErrorsToErrorVaults, mapWarningsToWarningVaults } from 'features/aave/helpers'
+import { useAutomationPrimaryButton } from 'features/aave/manage/helpers/use-automation-primary-button'
 import { getTriggerExecutionPrice } from 'features/aave/manage/services/calculations'
 import type {
   AutoSellTriggerAaveContext,
@@ -42,7 +43,7 @@ type AutoSellTriggerAaveContextWithPosition = AutoSellTriggerAaveContext & {
   position: PositionLike
 }
 
-interface AutoSellSidebarAaveVaultProps {
+export interface AutoSellSidebarAaveVaultProps {
   strategy: IStrategyConfig
   state: AutoSellTriggerAaveContextWithPosition
   isStateMatch: (state: BasicAutomationAaveState) => boolean
@@ -338,91 +339,6 @@ export function SideBarContent(props: AutoSellSidebarAaveVaultProps) {
   return <></>
 }
 
-export function usePrimaryButton(
-  props: AutoSellSidebarAaveVaultProps,
-): SidebarSectionFooterButtonSettings {
-  const { isStateMatch, canTransitWith } = props
-  const { t } = useTranslation()
-  const editingLabel =
-    props.state.action === TriggerAction.Add
-      ? t('automation.add-trigger', {
-          feature: t(sidebarAutomationFeatureCopyMap[props.state.feature]),
-        })
-      : t('automation.update-trigger', {
-          feature: t(sidebarAutomationFeatureCopyMap[props.state.feature]),
-        })
-
-  switch (true) {
-    case isStateMatch('idle'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {},
-        disabled: true,
-        label: editingLabel,
-        steps: [1, 3],
-      }
-    case isStateMatch('editing'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {
-          props.updateState({ type: 'REVIEW_TRANSACTION' })
-        },
-        disabled: !canTransitWith({ type: 'REVIEW_TRANSACTION' }),
-        label: editingLabel,
-        steps: [1, 3],
-      }
-    case isStateMatch('review'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {
-          props.updateState({ type: 'START_TRANSACTION' })
-        },
-        disabled: !canTransitWith({ type: 'START_TRANSACTION' }),
-        label: props.state.retryCount > 0 ? t('retry') : t('protection.confirm'),
-        steps: [2, 3],
-      }
-    case isStateMatch('remove'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {
-          props.updateState({ type: 'START_TRANSACTION' })
-        },
-        disabled: !canTransitWith({ type: 'START_TRANSACTION' }),
-        label:
-          props.state.retryCount > 0
-            ? t('retry')
-            : t('automation.cancel-trigger', {
-                feature: t(sidebarAutomationFeatureCopyMap[props.state.feature]),
-              }),
-      }
-    case isStateMatch('tx'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {},
-        disabled: true,
-        label: t('automation.setting', {
-          feature: t(sidebarAutomationFeatureCopyMap[props.state.feature]),
-        }),
-        steps: [3, 3],
-      }
-    case isStateMatch('txDone'):
-      return {
-        isLoading: props.state.isLoading,
-        action: () => {
-          props.updateState({ type: 'RESET' })
-        },
-        disabled: false,
-        label: t('finished'),
-      }
-  }
-  return {
-    isLoading: props.state.isLoading,
-    action: () => {},
-    disabled: true,
-    label: '',
-  }
-}
-
 export function useTextButton(
   props: AutoSellSidebarAaveVaultProps,
 ): SidebarSectionFooterButtonSettings | undefined {
@@ -461,7 +377,7 @@ export function AutoSellSidebarAaveVault(props: AutoSellSidebarAaveVaultProps) {
 
   const { strategy } = props
 
-  const primaryButton = usePrimaryButton(props)
+  const primaryButton = useAutomationPrimaryButton(props)
   const textButton = useTextButton(props)
 
   return (
